Minor fixes to the Vitest setup
