+++ .gitignore
@@ -41,17 +41,13 @@ jspm_packages
 .DS_Store
 
 /dist
-# Test Build Directories
-/dist_test
-/sample_react_project/
-/sample_custom_react_project/
 /keycloakify_starter_test/
+/sample_custom_react_project/
+/sample_react_project/
 /.yarn_home/
 
 .idea
 
-/keycloak_email
-/build_keycloak
 /src/login/i18n/baseMessages/
 /src/account/i18n/baseMessages/
