+++ vitest.config.ts
@@ -3,10 +3,10 @@ import { defineConfig } from "vite";
 import path from "path";
 
 export default defineConfig({
-    test: {
-        alias: {
+    "test": {
+        "alias": {
             "keycloakify": path.resolve(__dirname, "./src")
         },
-        watchExclude: ["**/node_modules/**", "**/dist/**", "**/sample_react_project/**", "**/sample_custom_react_project/**"]
+        "watchExclude": ["**/node_modules/**", "**/dist/**", "**/sample_react_project/**", "**/sample_custom_react_project/**"]
     }
 });
